+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Documentation for hipFFT is available at
 
 ### Changes
 
+* Compile with amdclang++ instead of hipcc for AMD backend; CUDA back-end still uses hipcc-nvcc.
 * Replace Boost Program Options with CLI11 as the command line parser for clients.
 
 ## hipFFT 1.0.14 for ROCm 6.1.0